big refactor - oop stuff + ui redesign first steps
