+++ KeyStash/Components/InfoButton.swift
@@ -4,23 +4,22 @@ struct InfoButton: View {
 	let label: String
 	let value: String
 	let onClick: () -> Void
-	
+
 	@State private var isHovering = false
 	@State private var clicked = false
-	
+
 	var body: some View {
 		Button(action: clickHandler, label: {
-			Image(systemName: clicked ? "checkmark.circle.fill" : "doc.on.doc.fill")
-				.foregroundStyle(.accent)
-				.opacity(isHovering ? 1 : 0.5)
-				.transition(.opacity)
-				.animation(.easeInOut(duration: 0.15), value: isHovering)
-				.contentTransition(.symbolEffect(.replace.magic(fallback: .upUp.byLayer), options: .nonRepeating))
-			VStack(alignment: .leading) {
-				Text(label)
-					.font(.caption)
-					.monospaced(false)
-				Text(value)
+			VStack(alignment: .trailing) {
+				if clicked {
+					Label("Copied", systemImage: "document.fill")
+						.transition(.scale)
+				} else {
+					Text(value)
+						.transition(.scale)
+						.truncationMode(.tail)
+						.lineLimit(isHovering ? 10 : 1)
+				}
 			}
 		})
 		.buttonStyle(.plain)
@@ -37,18 +36,22 @@ struct InfoButton: View {
 		})
 		.multilineTextAlignment(.leading)
 	}
-	
+
 	private func clickHandler() {
 		if clicked == false {
 			onClick()
-			clicked = true
-			
+			withAnimation {
+				clicked = true
+			}
+
 			DispatchQueue.main.asyncAfter(deadline: .now() + 1.0) {
 				yourFuncHere()
 			}
 
 			func yourFuncHere() {
-				clicked = false
+				withAnimation {
+					clicked = false
+				}
 			}
 		}
 	}
