+++ KeyStash/Views/LicenseInfoView/InfoRow.swift
@@ -2,18 +2,20 @@ import SwiftUI
 
 struct InfoRow: View {
 	@EnvironmentObject var appState: AppState
-	
+
 	var label: String
 	var value: String
-	
+
 	init(_ label: String, value: String) {
 		self.label = label
 		self.value = value
 	}
-	
+
 	var body: some View {
-		HStack(alignment: .top) {
-			if !value.isEmpty {
+		if !value.isEmpty {
+			HStack {
+				Text(label)
+				Spacer()
 				InfoButton(
 					label: label,
 					value: value,
@@ -22,15 +24,14 @@ struct InfoRow: View {
 					}
 				)
 				.contextMenu {
-					Button("Copy \"\(value)\"") {
+					Button("Copy \(label)") {
 						copyAction(value)
 					}
 				}
-				Spacer()
 			}
 		}
 	}
-	
+
 	private func copyAction(_ value: String) {
 		stringToClipboard(value: value)
 	}
