+++ KeyStash/Views/LicenseInfoView/_LicenseInfoView.swift
@@ -8,24 +8,19 @@ struct LicenseInfoView: View {
 	@Binding var selectedLicense: License
 
 	var body: some View {
-		ScrollView {
-			VStack {
-				LicenseHeader(selectedLicense)
-				LicenseDetails(selectedLicense)
+		LicenseDetails(selectedLicense)
+			.frame(maxWidth: .infinity)
+			.environmentObject(formState)
+			.toolbar {
+				ToolbarItem {
+					Spacer()
+				}
+				ToolbarItem(placement: .primaryAction) {
+					EditButton(selectedLicense)
+				}
 			}
-		}
-		.frame(maxWidth: .infinity)
-		.environmentObject(formState)
-		.toolbar {
-			ToolbarItem {
-				Spacer()
+			.sheet(isPresented: $appState.showEditAppSheet) {
+				EditLicenseView(isPresented: $appState.showEditAppSheet, license: selectedLicense)
 			}
-			ToolbarItem(placement: .cancellationAction) {
-				EditButton(selectedLicense)
-			}
-		}
-		.sheet(isPresented: $appState.showEditAppSheet) {
-			EditLicenseView(isPresented: $appState.showEditAppSheet, license: selectedLicense)
-		}
 	}
 }
